Tests: utils module introduced.
