+++ test/test_asgi_lifespan.py
@@ -3,7 +3,6 @@ from distutils.version import LooseVersion
 
 import pytest
 
-from conftest import public_dir
 from conftest import unit_stop
 from unit.applications.lang.python import TestApplicationPython
 from unit.option import option
