+++ test/test_proxy.py
@@ -6,9 +6,9 @@ import pytest
 
 from conftest import run_process
 from conftest import skip_alert
-from conftest import waitforsocket
 from unit.applications.lang.python import TestApplicationPython
 from unit.option import option
+from unit.utils import waitforsocket
 
 
 class TestProxy(TestApplicationPython):
