+++ test/test_static.py
@@ -3,9 +3,9 @@ import socket
 
 import pytest
 
-from conftest import waitforfiles
 from unit.applications.proto import TestApplicationProto
 from unit.option import option
+from unit.utils import waitforfiles
 
 
 class TestStatic(TestApplicationProto):
